fix dataFromJson function
