+++ src/main/java/hw6/entities/MetalsAndColorsData.java
@@ -1,23 +1,50 @@
 package hw6.entities;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 public class MetalsAndColorsData {
-    private Data data_1;
-    private Data data_2;
-    private Data data_3;
-    private Data data_4;
-    private Data data_5;
+   private String color;
+   private String metals;
+   private int[] summary;
+   private String[] vegetables;
+   private String[] elements;
 
-    public List<Data> getData(){
-        List<Data> data_list = new ArrayList<>();
-        data_list.add(data_1);
-        data_list.add(data_2);
-        data_list.add(data_3);
-        data_list.add(data_4);
-        data_list.add(data_5);
-        return data_list;
-    }
+   public String getColor() {
+       return color;
+   }
+
+   public String getMetals() {
+       return metals;
+   }
+
+   public int getSummaryOdd() {
+       return summary[0];
+   }
+
+   public int getSummaryEven() {
+       return summary[1];
+   }
+
+   public String[] getVegetables() {
+       return vegetables;
+   }
+
+   public String[] getElements() {
+       return elements;
+   }
+
+
+   public List<String> getListOfPageElements() {
+       List<String> result = new ArrayList<>();
+       result.add(color);
+       result.add(metals);
+       result.add(String.valueOf(summary[0] + summary[1]));
+       result.add(Arrays.toString(vegetables));
+       result.add((Arrays.toString(elements)));
+
+       return result;
+   }
 }
