+++ src/test/java/hw6/TestDataFromJson.java
@@ -1,14 +1,16 @@
 package hw6;
 
 import com.google.gson.Gson;
-import hw6.entities.Data;
+import com.google.gson.reflect.TypeToken;
 import hw6.entities.MetalsAndColorsData;
+//import hw6.entities.MetalsAndColorsData;
 import org.testng.annotations.DataProvider;
 
 import java.io.FileNotFoundException;
 import java.io.FileReader;
+import java.lang.reflect.Type;
 import java.util.Iterator;
-import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class TestDataFromJson {
@@ -16,8 +18,9 @@ public class TestDataFromJson {
 
     @DataProvider(name = "TestData")
     public Iterator<Object[]> getTestDataFromJson() throws FileNotFoundException {
-        MetalsAndColorsData data = new Gson().fromJson(new FileReader(JSON_PATH), MetalsAndColorsData.class);
-        List<Data> array_data = data.getData();
-        return  array_data.stream().map(Mdata -> new Object[]{Mdata}).collect(Collectors.toList()).iterator();
+        Type MetalsAndColorsMapType = new TypeToken<Map<String, MetalsAndColorsData>>() {}.getType();
+        Map<String, MetalsAndColorsData> data = new Gson().fromJson(new FileReader(JSON_PATH), MetalsAndColorsMapType);
+
+        return data.values().stream().map(data1 -> new Object[]{data1}).collect(Collectors.toList()).iterator();
     }
 }
